Identify the longest "good" stretch of trials (#6)
In go/no-go behavior, mice tend to lick exuberantly early in the
session, presumably because they are willing to see whether it
gets them water. Conversely, once satiated they tend to lick
sparsely. This aims to find a block of trials in the middle where
the fraction of lick trials matches the fraction of positive trials.